+++ packages/react-core/src/components/Tabs/Tabs.tsx
@@ -8,6 +8,7 @@ import AngleRightIcon from '@patternfly/react-icons/dist/js/icons/angle-right-ic
 import { getUniqueId, isElementInView, formatBreakpointMods } from '../../helpers/util';
 import { TabButton } from './TabButton';
 import { TabContent } from './TabContent';
+import { TabProps } from './Tab';
 import { getOUIAProps, OUIAProps, getDefaultOUIAId } from '../../helpers';
 
 export enum TabsComponent {
@@ -16,10 +17,12 @@ export enum TabsComponent {
 }
 
 export interface TabsProps extends Omit<React.HTMLProps<HTMLElement | HTMLDivElement>, 'onSelect'>, OUIAProps {
-  /** Content rendered inside the tabs component. */
+  /** Content rendered inside the tabs component. Must be React.ReactElement<TabProps>[] */
   children: React.ReactNode;
   /** Additional classes added to the tabs */
   className?: string;
+  /** Tabs background color variant */
+  variant?: 'default' | 'light300';
   /** The index of the active tab */
   activeKey?: number | string;
   /** Callback to handle tab selection */
@@ -57,6 +60,22 @@ export interface TabsProps extends Omit<React.HTMLProps<HTMLElement | HTMLDivEle
   };
 }
 
+export interface TabsContextProps {
+  variant: 'default' | 'light300';
+}
+
+const TabsContext = React.createContext<TabsContextProps>({
+  variant: 'default'
+});
+
+export const TabsContextProvider = TabsContext.Provider;
+export const TabsContextConsumer = TabsContext.Consumer;
+
+const variantStyle = {
+  default: '',
+  light300: styles.modifiers.colorSchemeLight_300
+};
+
 interface TabsState {
   showScrollButtons: boolean;
   disableLeftScrollButton: boolean;
@@ -91,12 +110,13 @@ export class Tabs extends React.Component<TabsProps, TabsState> {
     component: TabsComponent.div,
     mountOnEnter: false,
     unmountOnExit: false,
-    ouiaSafe: true
+    ouiaSafe: true,
+    variant: 'default'
   };
 
   handleTabClick(
     event: React.MouseEvent<HTMLElement, MouseEvent>,
-    eventKey: number,
+    eventKey: number | string,
     tabContentRef: React.RefObject<any>,
     mountOnEnter: boolean
   ) {
@@ -104,7 +124,8 @@ export class Tabs extends React.Component<TabsProps, TabsState> {
     this.props.onSelect(event, eventKey);
     // process any tab content sections outside of the component
     if (tabContentRef) {
-      React.Children.toArray<any>(this.props.children)
+      React.Children.toArray(this.props.children)
+        .map(child => child as React.ReactElement<TabProps>)
         .filter(child => child.props && child.props.tabContentRef && child.props.tabContentRef.current)
         .forEach(child => (child.props.tabContentRef.current.hidden = true));
       // most recently selected tabContent
@@ -223,15 +244,19 @@ export class Tabs extends React.Component<TabsProps, TabsState> {
       mountOnEnter,
       unmountOnExit,
       inset,
+      variant,
       ...props
     } = this.props;
     const { showScrollButtons, disableLeftScrollButton, disableRightScrollButton, shownKeys } = this.state;
+    const filteredChildren = (React.Children.toArray(children) as React.ReactElement<TabProps>[])
+      .filter(Boolean)
+      .filter(child => !child.props.isHidden);
 
     const uniqueId = id || getUniqueId();
     const Component: any = component === TabsComponent.nav ? 'nav' : 'div';
 
     return (
-      <React.Fragment>
+      <TabsContextProvider value={{ variant }}>
         <Component
           aria-label={ariaLabel}
           className={css(
@@ -242,6 +267,7 @@ export class Tabs extends React.Component<TabsProps, TabsState> {
             isBox && styles.modifiers.box,
             showScrollButtons && !isVertical && styles.modifiers.scrollable,
             formatBreakpointMods(inset, styles),
+            variantStyle[variant],
             className
           )}
           {...getOUIAProps(Tabs.displayName, ouiaId !== undefined ? ouiaId : this.state.ouiaStateId, ouiaSafe)}
@@ -258,45 +284,41 @@ export class Tabs extends React.Component<TabsProps, TabsState> {
             <AngleLeftIcon />
           </button>
           <ul className={css(styles.tabsList)} ref={this.tabList} onScroll={this.handleScrollButtons}>
-            {React.Children.toArray<any>(children)
-              .filter(Boolean)
-              .map((child, index) => {
-                const {
-                  title,
-                  eventKey,
-                  tabContentRef,
-                  id: childId,
-                  tabContentId,
-                  isHidden = false,
-                  className: childClassName = '',
-                  ouiaId: childOuiaId,
-                  ...rest
-                } = child.props;
-                let ariaControls = tabContentId
-                  ? `${tabContentId}`
-                  : `pf-tab-section-${eventKey}-${childId || uniqueId}`;
-                if ((mountOnEnter || unmountOnExit) && eventKey !== activeKey) {
-                  ariaControls = undefined;
-                }
-                return isHidden ? null : (
-                  <li
-                    key={index}
-                    className={css(styles.tabsItem, eventKey === activeKey && styles.modifiers.current, childClassName)}
+            {filteredChildren.map((child, index) => {
+              const {
+                title,
+                eventKey,
+                tabContentRef,
+                id: childId,
+                tabContentId,
+                isHidden,
+                className: childClassName = '',
+                ouiaId: childOuiaId,
+                ...rest
+              } = child.props;
+              let ariaControls = tabContentId ? `${tabContentId}` : `pf-tab-section-${eventKey}-${childId || uniqueId}`;
+              if ((mountOnEnter || unmountOnExit) && eventKey !== activeKey) {
+                ariaControls = undefined;
+              }
+              return (
+                <li
+                  key={index}
+                  className={css(styles.tabsItem, eventKey === activeKey && styles.modifiers.current, childClassName)}
+                >
+                  <TabButton
+                    className={css(styles.tabsLink)}
+                    onClick={(event: any) => this.handleTabClick(event, eventKey, tabContentRef, mountOnEnter)}
+                    id={`pf-tab-${eventKey}-${childId || uniqueId}`}
+                    aria-controls={ariaControls}
+                    tabContentRef={tabContentRef}
+                    ouiaId={childOuiaId}
+                    {...rest}
                   >
-                    <TabButton
-                      className={css(styles.tabsLink)}
-                      onClick={(event: any) => this.handleTabClick(event, eventKey, tabContentRef, mountOnEnter)}
-                      id={`pf-tab-${eventKey}-${childId || uniqueId}`}
-                      aria-controls={ariaControls}
-                      tabContentRef={tabContentRef}
-                      ouiaId={childOuiaId}
-                      {...rest}
-                    >
-                      {title}
-                    </TabButton>
-                  </li>
-                );
-              })}
+                    {title}
+                  </TabButton>
+                </li>
+              );
+            })}
           </ul>
           <button
             className={css(styles.tabsScrollButton, isSecondary && buttonStyles.modifiers.secondary)}
@@ -308,10 +330,9 @@ export class Tabs extends React.Component<TabsProps, TabsState> {
             <AngleRightIcon />
           </button>
         </Component>
-        {React.Children.toArray<any>(children)
+        {filteredChildren
           .filter(
             child =>
-              child &&
               child.props.children &&
               !(unmountOnExit && child.props.eventKey !== activeKey) &&
               !(mountOnEnter && shownKeys.indexOf(child.props.eventKey) === -1)
@@ -325,7 +346,7 @@ export class Tabs extends React.Component<TabsProps, TabsState> {
               ouiaId={child.props.ouiaId}
             />
           ))}
-      </React.Fragment>
+      </TabsContextProvider>
     );
   }
 }