+++ packages/react-docs/patternfly-docs.config.js
@@ -1,0 +1,18 @@
+// This module is shared between NodeJS and babelled ES5
+module.exports = {
+  pathPrefix: '',
+  hasGdprBanner: false,
+  hasFooter: false,
+  hasVersionSwitcher: false,
+  hasDesignGuidelines: false,
+  sideNavItems: [
+    { section: 'get-started' },
+    { section: 'charts' },
+    { section: 'components' },
+    { section: 'demos' },
+    { section: 'extensions' },
+    { section: 'layouts' }
+  ],
+  topNavItems: [{ text: 'Icons', path: '/icons' }],
+  port: 8002
+};