+++ packages/react-integration/demo-app-ts/src/components/demos/DonutChartDemo/DonutSmallDemo.tsx
@@ -12,7 +12,11 @@ export class DonutSmallDemo extends React.Component {
       <div>
         <div style={{ height: '150px', width: '150px', paddingTop: '50px', paddingLeft: '50px' }}>
           <ChartDonut
-            data={[{ x: 'Cats', y: 35 }, { x: 'Dogs', y: 55 }, { x: 'Birds', y: 10 }]}
+            data={[
+              { x: 'Cats', y: 35 },
+              { x: 'Dogs', y: 55 },
+              { x: 'Birds', y: 10 }
+            ]}
             height={150}
             labels={({ datum }) => `${datum.x}: ${datum.y}%`}
             subTitle="Pets"