+++ packages/react-styles/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.7.6](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-styles@4.7.5...@patternfly/react-styles@4.7.6) (2020-09-28)
+
+**Note:** Version bump only for package @patternfly/react-styles
+
+
+
+
+
 ## [4.7.5](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-styles@4.7.4...@patternfly/react-styles@4.7.5) (2020-09-17)
 
 **Note:** Version bump only for package @patternfly/react-styles